backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
